+++ ellip-dev-utils/src/wrapper.rs
@@ -5,18 +5,18 @@
 
 #[macro_export]
 macro_rules! func_wrapper {
-    ($func:expr, 1) => {
-        fn wrapped_func(args: &Vec<f64>) -> f64 {
+    ($func:expr, $t: ty, 1) => {
+        fn wrapped_func(args: &Vec<$t>) -> $t {
             $func(args[0]).unwrap()
         }
     };
-    ($func:expr, 2) => {
-        fn wrapped_func(args: &Vec<f64>) -> f64 {
+    ($func:expr, $t: ty, 2) => {
+        fn wrapped_func(args: &Vec<$t>) -> $t {
             $func(args[0], args[1]).unwrap()
         }
     };
-    ($func:expr, 3) => {
-        fn wrapped_func(args: &Vec<f64>) -> f64 {
+    ($func:expr, $t: ty, 3) => {
+        fn wrapped_func(args: &Vec<$t>) -> $t {
             let ans = $func(args[0], args[1], args[2]).unwrap_or_else(|e| {
                 println!("{}: {}, {}, {}", e, args[0], args[1], args[2]);
                 panic!()
@@ -28,13 +28,13 @@ macro_rules! func_wrapper {
             ans
         }
     };
-    ($func:expr, 4) => {
-        fn wrapped_func(args: &Vec<f64>) -> f64 {
+    ($func:expr, $t: ty, 4) => {
+        fn wrapped_func(args: &Vec<$t>) -> $t {
             $func(args[0], args[1], args[2], args[3]).unwrap()
         }
     };
-    ($func:expr, $_:expr) => {
-        fn wrapped_func(_args: &Vec<f64>) -> f64 {
+    ($func:expr, $t: ty, $_:expr) => {
+        fn wrapped_func(_args: &Vec<$t>) -> $t {
             panic!("Unsupported number of arguments")
         }
     };