+++ examples/generate_test_summary.rs
@@ -7,14 +7,14 @@ use ellip::*;
 use ellip_dev_utils::{env::get_env, get_summary_entry, test_report::generate_summary_table};
 
 fn main() {
-    let [rust_version, platform, ellip_version] = get_env();
+    let env = get_env();
     let template = std::fs::read_to_string("examples/README_template.md")
         .expect("Cannot read README template");
 
     let summary_section = [
         &format!(
-            "Generated on {} rustc {} using ellip v{} at `f64` precision (ε≈2.22e-16).\n",
-            platform, rust_version, ellip_version
+            "Benchmark on {} running {} rustc {} using ellip v{} at `f64` precision (ε≈2.22e-16).\n",
+            env.cpu, env.platform, env.rust_version, env.ellip_version
         ),
         "### Legendre's Elliptic Integrals",
         &generate_summary_table(&[
@@ -24,6 +24,7 @@ fn main() {
             get_summary_entry!("legendre", "ellipeinc", ellipeinc, 2),
             get_summary_entry!("legendre", "ellippi", ellippi, 2),
             get_summary_entry!("legendre", "ellippiinc", ellippiinc, 3),
+            get_summary_entry! {"legendre", "ellippiinc_bulirsch", ellippiinc_bulirsch, 3, "ellippiinc"},
             get_summary_entry!("legendre", "ellipd", ellipd, 1),
             get_summary_entry!("legendre", "ellipdinc", ellipdinc, 2),
         ]),
@@ -45,6 +46,12 @@ fn main() {
             get_summary_entry!("carlson", "elliprj", elliprj, 4),
             get_summary_entry!("carlson", "elliprc", elliprc, 2),
             get_summary_entry!("carlson", "elliprd", elliprd, 3),
+        ]),
+        "",
+        "### Miscellaneous Functions",
+        &generate_summary_table(&[
+            get_summary_entry!("misc", "jacobi_zeta", jacobi_zeta, 2),
+            get_summary_entry!("misc", "heuman_lambda", heuman_lambda, 2),
         ]),
     ];