+++ examples/plot-graph/src/bin/cel1.rs
@@ -47,7 +47,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
             )
             .annotations(vec![Annotation::new()
             .text(format!(
-                "Generated using the function <a href=\"https://docs.rs/ellip/latest/ellip/bulirsch/fn.cel1.html\" target=\"_blank\">cel1</a> from <a href=\"https://crates.io/crates/ellip\" target=\"_blank\">ellip</a> v{}",
+                "Generated using <a href=\"https://docs.rs/ellip/latest/ellip/bulirsch/fn.cel1.html\" target=\"_blank\">cel1</a> from <a href=\"https://crates.io/crates/ellip\" target=\"_blank\">ellip</a> v{}",
                 ellip_version()
             ))
                 .x_ref("paper")