+++ examples/plot-graph/src/bin/el3.rs
@@ -120,7 +120,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
              ).legend(Legend::new().y_anchor(Anchor::Middle).y(0.5))
              .annotations(vec![Annotation::new()
              .text(format!(
-                 "Generated using the function <a href=\"https://docs.rs/ellip/latest/ellip/bulirsch/fn.el3.html\" target=\"_blank\">el3</a> from <a href=\"https://crates.io/crates/ellip\" target=\"_blank\">ellip</a> v{}",
+                 "Generated using <a href=\"https://docs.rs/ellip/latest/ellip/bulirsch/fn.el3.html\" target=\"_blank\">el3</a> from <a href=\"https://crates.io/crates/ellip\" target=\"_blank\">ellip</a> v{}",
                  ellip_version()
              ))
                  .x_ref("paper")