+++ examples/plot-graph/src/bin/ellippiinc-3d.rs
@@ -6,9 +6,9 @@
 use ellip::ellippiinc;
 use ellip_plot_graph::*;
 use plotly::{
-    Layout, Plot, Surface,
     common::{ColorScale, ColorScalePalette},
     layout::{Annotation, AspectRatio, Axis, LayoutScene},
+    Layout, Plot, Surface,
 };
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -141,7 +141,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
              )
              .annotations(vec![Annotation::new()
              .text(format!(
-                 "Generated using the function <a href=\"https://docs.rs/ellip/latest/ellip/legendre/fn.ellippiinc.html\" target=\"_blank\">ellippiinc</a> from <a href=\"https://crates.io/crates/ellip\" target=\"_blank\">ellip</a> v{}",
+                 "Generated using <a href=\"https://docs.rs/ellip/latest/ellip/legendre/fn.ellippiinc.html\" target=\"_blank\">ellippiinc</a> from <a href=\"https://crates.io/crates/ellip\" target=\"_blank\">ellip</a> v{}",
                  ellip_version()
              ))
              .x_ref("paper")