+++ src/legendre/ellipeinc.rs
@@ -83,8 +83,14 @@ pub fn ellipeinc<T: Float>(phi: T, m: T) -> Result<T, StrErr> {
     Err("ellipeinc: Unexpected error.")
 }
 
-#[numeric_literals::replace_float_literals(T::from(literal).unwrap())]
+/// Unsafe version of [ellipeinc].
+/// <div class="warning">⚠️ Unstable feature. May subject to changes.</div>
+///
+/// Undefined behavior with invalid arguments and edge cases.
+/// # Known Invalid Cases
+/// - NAN arguments
 #[inline]
+#[numeric_literals::replace_float_literals(T::from(literal).unwrap())]
 pub fn ellipeinc_unchecked<T: Float>(phi: T, m: T) -> Result<T, StrErr> {
     if phi == 0.0 {
         return Ok(0.0);