+++ src/legendre/mod.rs
@@ -21,6 +21,7 @@ pub use ellipeinc::ellipeinc;
 pub use ellipf::ellipf;
 pub use ellipk::ellipk;
 pub use ellippi::ellippi;
-pub use ellippiinc::ellippiinc;
+pub use ellippiinc::{ellippiinc, ellippiinc_bulirsch, ellippiinc_bulirsch_with_const};
 
-pub(crate) use ellipeinc::ellipeinc_unchecked;
+#[cfg(feature = "unstable")]
+pub use ellippi::ellippi_unchecked;